+++ prefs.py
@@ -34,6 +34,7 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
         default = 8.0, min = 0.001, max = 20000, soft_min = 0.01, soft_max = 4000, step = 10, precision = 2)
     wasdTime: bpy.props.FloatProperty(name = "WASD Acceleration Time", description = "Time until top speed is reached when using WASD keys to move", \
         default = 0.2, min = 0.0, max = 4.0, soft_min = 0.0, soft_max = 1000, step = 1, precision = 2)
+    wasdGlobalZ: bpy.props.BoolProperty(name = "Use Global Z", description = "When checked, makes WASD up/down movement aligned to global Z-axis instead of view Z-axis", default = False)
 
     mouseWheelActionItems = [\
         ("moveZ", "Move forward/backwards", "Move forward/backwards", "NONE", 0), \
@@ -137,6 +138,9 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
         row = box.row()
         row.prop(self, "wasdTopSpeed")
         row.prop(self, "wasdTime")
+
+        row = box.row()
+        row.prop(self, "wasdGlobalZ")
 
     def drawKeyBindPrefs(self, layout: bpy.types.UILayout):
         layout.label(text = "Key Bindings:", translate = False)